+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/eventpublishing/configuration/EventLinkingWorkerConfig.java
@@ -28,6 +28,10 @@ public class EventLinkingWorkerConfig {
     private String insertEventIntoPublishedEventTable;
 
     @Inject
+    @Value(key = "event.linking.worker.batch.size", defaultValue = "10")
+    private String batchSize;
+
+    @Inject
     @Value(key = "event.linking.worker.transaction.timeout.seconds", defaultValue = DEFAULT_TIMEOUT_SECONDS)
     private String transactionTimeoutSeconds;
 
@@ -49,6 +53,10 @@ public class EventLinkingWorkerConfig {
 
     public boolean shouldAlsoInsertEventIntoPublishedEventTable() {
         return parseBoolean(insertEventIntoPublishedEventTable);
+    }
+
+    public int getBatchSize() {
+        return parseInt(batchSize);
     }
 
     public int getTransactionTimeoutSeconds() {