+++ event-sourcing/event-publisher/published-event-processor/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/jdbc/LinkedEventData.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.services.eventsourcing.publishedevent.jdbc;
+
+import java.util.UUID;
+
+public record LinkedEventData(UUID eventId, long eventNumber, long previousEventNumber) {
+}