+++ event-sourcing/event-publisher/published-event-processor/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/jdbc/AdvisoryLockDataAccessIT.java
@@ -2,42 +2,26 @@ package uk.gov.justice.services.eventsourcing.publishedevent.jdbc;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.when;
 
-import uk.gov.justice.services.eventsourcing.source.core.EventStoreDataSourceProvider;
 import uk.gov.justice.services.test.utils.persistence.FrameworkTestDataSourceFactory;
-
-import javax.sql.DataSource;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 public class AdvisoryLockDataAccessIT {
 
-    @Spy
-    private EventStoreDataSourceProvider eventStoreDataSourceProvider;
-
     @InjectMocks
     private AdvisoryLockDataAccess advisoryLockDataAccess;
 
     @Test
-    public void shouldObtainBlockingTransactionalLock() throws Exception {
+    public void shouldTryToObtainNonBlockingTransactionalLockUsingProvidedConnection() throws Exception {
+        final javax.sql.DataSource eventStoreDataSource = new FrameworkTestDataSourceFactory().createEventStoreDataSource();
 
-        final DataSource eventStoreDataSource = new FrameworkTestDataSourceFactory().createEventStoreDataSource();
-        when(eventStoreDataSourceProvider.getDefaultDataSource()).thenReturn(eventStoreDataSource);
-
-        advisoryLockDataAccess.obtainBlockingTransactionLevelAdvisoryLock(23L);
-    }
-
-    @Test
-    public void shouldTryToObtainNonBlockingTransactionalLock() throws Exception {
-        final DataSource eventStoreDataSource = new FrameworkTestDataSourceFactory().createEventStoreDataSource();
-        when(eventStoreDataSourceProvider.getDefaultDataSource()).thenReturn(eventStoreDataSource);
-
-        assertThat(advisoryLockDataAccess.tryNonBlockingTransactionLevelAdvisoryLock(23L), is(true));
+        try (final java.sql.Connection connection = eventStoreDataSource.getConnection()) {
+            assertThat(advisoryLockDataAccess.tryNonBlockingTransactionLevelAdvisoryLock(connection, 42L), is(true));
+        }
     }
 }