+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased]
+- Batch event linking: `EventNumberLinker` now links N events per JTA transaction using JDBC `executeBatch()`, 
+  configurable via JNDI `event.linking.worker.batch.size` (default 10)
 
 # [17.104.0] - 2025-12-16
 ### Added